refactor: reorganize tests into src/ directory with logical modules
- Move tests from tests/ directory to src/ as unit tests
- Split tests into 5 focused modules:
  * server_tests.rs - server initialization and info
  * client_tests.rs - client initialization and config
  * api_tests.rs - HTTP API interactions and functionality
  * formatting_tests.rs - search result formatting
  * security_tests.rs - TLS and security flag handling
- Update lib.rs with test module declarations
- Remove old tests/ directory structure
- All 17 tests continue to pass
