+++ src/lib.rs
@@ -5,3 +5,14 @@ pub mod server;
 pub use client::*;
 pub use formatting::*;
 pub use server::*;
+
+#[cfg(test)]
+mod api_tests;
+#[cfg(test)]
+mod client_tests;
+#[cfg(test)]
+mod formatting_tests;
+#[cfg(test)]
+mod security_tests;
+#[cfg(test)]
+mod server_tests;
